# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Oculide Core REST API
  description: >
    Hệ thống API RESTful và WebSocket phục vụ Giám sát phòng thực hành lập trình & Gác thi bằng AI (Oculide v1).
    Được thiết kế theo tiêu chuẩn IEEE 830-1998 Specification.
  version: 1.0.0
  contact:
    name: Oculide Development Team
    email: lean80469@gmail.com
servers:
  - url: https://api.oculide.id.vn/api/v1
    description: Production API Server
  - url: http://localhost:8000/api/v1
    description: Local Development Server
paths:
...
  /submissions:
    post:
      tags: [Submissions]
      summary: Nộp bài làm chính thức để chấm điểm tự động bằng Docker Sandbox
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmissionCreateRequest'
      responses:
        '201':
          description: Bài nộp đã được tiếp nhận và đưa vào hàng đợi chấm
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionResponse'
  /submissions/run:
    post:
      tags: [Submissions]
      summary: Chạy thử nghiệm code với Custom Input (không lưu điểm)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CodeRunRequest'
      responses:
        '200':
          description: Kết quả thực thi trong Sandbox
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodeRunResponse'
  /submissions/{submission_id}:
    get:
      tags: [Submissions]
      summary: Lấy chi tiết bài nộp và trạng thái chấm
      security:
        - BearerAuth: []
      parameters:
        - name: submission_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Chi tiết bài nộp
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionResponse'
  /submissions/{submission_id}/results:
    get:
      tags: [Submissions]
      summary: Lấy kết quả chi tiết từng test case của bài nộp
      security:
        - BearerAuth: []
      parameters:
        - name: submission_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Danh sách kết quả test cases

  # ==========================================
  # SESSIONS
  # ==========================================
...
  /violations:
    post:
      tags: [Violations]
      summary: Ghi nhận sự kiện vi phạm
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ViolationCreateRequest'
      responses:
        '201':
          description: Ghi nhận vi phạm thành công
  /violations/analyze:
    post:
      tags: [Violations]
      summary: Gửi ảnh chụp webcam vào AI Worker để phân tích bất thường
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SnapshotAnalyzeRequest'
      responses:
        '202':
          description: Đã tiếp nhận ảnh vào hàng đợi phân tích
  /violations/{violation_id}/review:
    patch:
      tags: [Violations]
      summary: Giảng viên duyệt hoặc bỏ qua cảnh báo vi phạm
      security:
        - BearerAuth: []
      parameters:
        - name: violation_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Đã cập nhật trạng thái duyệt

  # ==========================================
  # LIVEKIT STREAMING
  # ==========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    SubmissionCreateRequest:
      type: object
      required: [room_id, question_id, code_content, language]
      properties:
        room_id:
          type: integer
        question_id:
          type: integer
        code_content:
          type: string
        language:
          type: string
          enum: [python, cpp, java, c]
        attempt_number:
          type: integer
          default: 1
    SubmissionResponse:
      type: object
      properties:
        submission_id:
          type: integer
        room_id:
          type: integer
        student_id:
          type: integer
        question_id:
          type: integer
        attempt_number:
          type: integer
        code_content:
          type: string
        language:
          type: string
        submitted_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, grading, completed, failed, saved]
        total_points:
          type: number
          nullable: true
        max_points:
          type: number
          nullable: true
    CodeRunRequest:
      type: object
      required: [code_content, language]
      properties:
        code_content:
          type: string
        language:
          type: string
        custom_input:
          type: string
          default: ''
    CodeRunResponse:
      type: object
      properties:
        stdout:
          type: string
        error:
          type: string
          nullable: true
        execution_ms:
          type: integer
...
    ViolationCreateRequest:
      type: object
      required: [session_id, violation_type]
      properties:
        session_id:
          type: integer
        violation_type:
          type: string
          enum: [tab_switch, copy_paste, no_face_detected, multiple_faces, phone_detected, suspicious_object, fullscreen_exit, camera_blocked, time_exceeded, inactive_30s, multiple_people]
        severity:
          type: string
          enum: [low, medium, high, critical]
          default: warning
        description:
          type: string
          nullable: true
        snapshot_url:
          type: string
          nullable: true
    SnapshotAnalyzeRequest:
      type: object
      required: [session_id, snapshot_data]
      properties:
        session_id:
          type: integer
        snapshot_data:
          type: string
          description: Base64-encoded webcam frame JPEG/PNG
...